Add draft menu code for some help
I'm having an issue and I can't figure out what it is. I'm trying to get it so that, when you click the labels (for the sizes), it displays the number counter for the user to select their quantity. But I can't get it to work, and I'm not sure what's wrong. The error that pops up is that there's a missing parenthesis, but I don't see one. Could either of you look and see if you can figure it out?

I'm going to keep trying tomorrow. Of course, maybe I should abandon the JSON idea, not sure if that's causing the issues.

I added the images that I'm using for testing. I plan to use the actual images soon. And I haven't set it up with the starter page yet.